Move interpretUserArgs to its own module
